+++ src/osc/Bundler/ChunkedImageBundler.hpp
@@ -7,7 +7,7 @@ class ChunkedManager;
 
 struct ChunkedImageBundler : ChunkedSendBundler {
   ChunkedImageBundler(
-    int32_t chunkedSendId,
+    int64_t chunkedSendId,
     int32_t chunkNum,
     int32_t numChunks,
     int32_t chunkSize,
