+++ src/osc/ChunkedManager.hpp
@@ -12,25 +12,25 @@ struct ChunkedManager {
   ~ChunkedManager();
 
   void add(ChunkedSend* chunked, bool deferIfAlreadyQueued = false);
-  void ack(int32_t id, int32_t chunkNum);
+  void ack(int64_t id, int32_t chunkNum);
 
-  bool isProcessing(int32_t id);
-  void processChunked(int32_t id);
+  bool isProcessing(int64_t id);
+  void processChunked(int64_t id);
 
   // used by bundlers. returns null if not found.
-  ChunkedSend* findChunked(int32_t id);
+  ChunkedSend* findChunked(int64_t id);
 
 private:
   OSCctrlWidget* ctrl{NULL};
   OscSender* osctx{NULL};
 
-  std::map<int32_t, std::unique_ptr<ChunkedSend>> chunkedSends;
-  bool chunkedExists(int32_t id);
+  std::map<int64_t, std::unique_ptr<ChunkedSend>> chunkedSends;
+  bool chunkedExists(int64_t id);
 
   void defer(ChunkedSend* chunked);
-  std::map<int32_t, ChunkedSend*> deferredSends;
-  bool deferredExists(int32_t id);
+  std::map<int64_t, ChunkedSend*> deferredSends;
+  bool deferredExists(int64_t id);
 
   // used internally, asserts chunked exists
-  ChunkedSend* getChunked(int32_t id);
+  ChunkedSend* getChunked(int64_t id);
 };
